examples/docker-kubernetes/services/control-plane/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "control-plane"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-agents = {git = "git@github.com:run-llama/llama-agents.git#nerdai/k8s"}
llama-index-llms-openai = "^0.1.23"
llama-index-embeddings-openai = "^0.1.10"
nest-asyncio = "^1.6.0"
llama-index-agent-openai = "^0.2.7"
